Phew! Major refactor
This isn't complete yet (need to handle renewal/rebind to get to parity)
bu this is a massive refactoring of the state machine to use a generator.
It obviously still needs the tests to be refactored.

The major goal is to increase readability by not having to carry around a bunch of
info across state machine <-> client boundaries, by relying on saving state in the
continuation. Since Racket generators are stackful, we can stack and nest yields,
and even persist state within various functions. I am hoping this will lead to much
cleaner code.