pass retry-state to requesting-state
right now, to-requesting-state mostly exists to better stick to the RFC,
where the first outgoing DHCPREQUEST is sent before the state is truly entered.

It also allows sticking to the "double-curried function" form for each state that I'm currently trying out.

Actually it could be trivially inlined.

Any way, because that initial outgoing packet needs to know the retry-determined deadline, do this shuffle.